pkg: add new `datasizes` package that adds constants and parsing
This commit extracts the existing `common.DataSizeToUint64` and
the `common.{Ki,Gi,Mi,...}` and similar data size constants into
a new package `datasizes` that will be shared with other packages
like `blueprints`.
